Add OpenTelemetry and Review Stack support
This PR depends on on [zephyr-k8s#8](https://github.com/pulumi/zephyr-k8s/pull/8), which installs
the AWS Distro for OpenTelemetry add-on in the cluster, which itself includes the [OpenTelemetry
Operator](https://github.com/open-telemetry/opentelemetry-operator) and deploys a shared
OpenTelemetry Collector.

To enable OpenTelemetry, we add an `Instrumentation` custom resource to each namespace, making it
the default for that namespace, and configuring pods in that namespace to send traces to the
OpenTelemetry Collector.

To enable review stacks, we remove the hardcoded namespace names and instead allow auto-naming to
give them random suffixes. For each of the cross-namespace service names, the namespace name is
interpolated into the environment variable via `ConfigMap` entries. This allows multiple deployments
of this stack to the same cluster, and each deployment can be uniquely identified in traces via
labeling in OpenTelemetry.

Footnotes
---------

**OpenTelemetry**

The OpenTelemetry operator, installed in the zephyr-k8s PR above, installs a
mutating admission webhook. That webhook's behavior depends on the
`Instrumentation` resource, so we may want to make any deployments or
statefulsets - any resource that creates pods - depend on the `Instrumentation`

**Review stacks**

The autonaming results in each namespace being given a random - not very useful for operators -
name. It might be nice to use the stack name some how, which includes the PR number, or to generate
a consistent **prefix** for each namespace instead.